pyproject.toml
[project]
name = "miso"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=4.8.1",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]
